Replaced PngWriter with StbImageWriteSharp
